replaced exceptions with console output

+++ src/main/java/school/faang/bjs245208/SampleData.java
@@ -65,6 +65,10 @@ public class SampleData {
 
     public static void removeStudent(StudentDatabase db) {
         Student student = db.findStudentByName("Student3");
+        if (student == null) {
+            System.err.println("Student does not exist!");
+            return;
+        }
         db.removeStudent(student.getId());
     }
